HS-34-01: split web/routes/dictation.py into routes/dictation/ sub-package
The 1,607-line single build_dictation_router(ctx) becomes a routes/dictation/
sub-package by domain, behind a STABLE build_dictation_router(ctx) (routes/
__init__.py unchanged):

- intents.py       /api/intents/*
- agent.py         /api/dictation/project-context, agent-context*, agent-hooks
- project_docs.py  /api/dictation/project-hs, project-doc-suggestion*
- blocks.py        /api/dictation/block-templates, blocks* (+ from-template)
- kb.py            /api/dictation/project-kb*
- pipeline.py      /api/dictation/readiness, dry-run
- _helpers.py      ctx-free shared helpers + starter templates + dry-run executor

Behavior-preserving: handlers/helpers moved verbatim (only ctx. target +
one extra relative-import dot); absolute paths + include_router keep the route
table identical (26 routes, hash 0a0b26562cf25a36 unchanged). The shared in-memory
project-doc-suggestion store (dry-run/from-template detect; project-doc-suggestion
GET reads) is created once in build_dictation_router and threaded to the three
groups that touch it, preserving the per-app closure lifetime.

New tests/unit/test_dictation_routes_split.py locks the exact (path, method) set
(Phase 34's shared route-table invariant). Suite green 1954/15; package ruff +
F821 clean.
